Add animations to santa!
