+++ client/src/app/game/game.ts
@@ -4,198 +4,8 @@ import { getConnectionInfo } from "@/api/room";
 import { Score } from "./page";
 import { MutableRefObject } from "react";
 import { getNickname, getSantaColor } from "@/lib/utils";
-import { SantaColor } from "@/lib/player-options";
 import { UP, DOWN, LEFT, RIGHT, NONE, MoveDirection } from "@common/input";
-
-const SANTA_SPRITE = "/santa-sprite.png";
-const SANTA_SPRITE_LEFT = "/santa-sprite-left.png";
-
-const SantaColorToSpriteRow = new Map<SantaColor, number>([
-  ["Red", 0],
-  ["Blue", 1],
-  ["Green", 2],
-  ["Yellow", 3],
-  ["Purple", 4],
-  ["Teal", 5],
-]);
-
-type Vec2d = {
-  x: number;
-  y: number;
-};
-
-class Sprite {
-  image: HTMLImageElement; // The actual sprite sheet
-  spriteDims: Vec2d; // The dimensions of the sprite sheet (width, height)
-  frameDims: Vec2d; // The dimensions of a single frame (width, height)
-  frameIndex: number; // The index of the current frame
-  frameMap: Map<number, Vec2d>; // A map of frame index to coords in sheet
-  position: Vec2d; // The position of the sprite on the canvas
-
-  constructor(
-    image: HTMLImageElement,
-    spriteDims: Vec2d,
-    frameDims: Vec2d,
-    startIndex: number,
-    startingPosition: Vec2d
-  ) {
-    this.image = image;
-    this.spriteDims = spriteDims;
-    this.frameDims = frameDims;
-    this.frameIndex = startIndex;
-    this.position = startingPosition ?? { x: 0, y: 0 };
-    this.frameMap = new Map<number, Vec2d>();
-  }
-
-  ij_to_flat_index(row: number, col: number) {
-    return row * this.spriteDims.x + col;
-  }
-
-  buildFrameMap() {
-    for (let c = 0; c < this.spriteDims.x; c++) {
-      for (let r = 0; r < this.spriteDims.y; r++) {
-        this.frameMap.set(this.ij_to_flat_index(r, c), {
-          x: c * this.frameDims.x,
-          y: r * this.frameDims.y,
-        });
-      }
-    }
-  }
-
-  draw(ctx: CanvasRenderingContext2D, x: number, y: number, frame: number) {
-    const theFrame = this.frameMap.get(frame);
-    if (!theFrame) {
-      throw new Error(`Could not find frame ${frame}`);
-    }
-    ctx.drawImage(
-      this.image,
-      theFrame.x,
-      theFrame.y,
-      this.frameDims.x,
-      this.frameDims.y,
-      x,
-      y,
-      this.frameDims.x,
-      this.frameDims.y
-    );
-  }
-}
-
-class Player {
-  id: string;
-  x: number;
-  nickname: string;
-  santaColor: SantaColor;
-  y: number;
-  isLeft: boolean;
-  kills: number;
-  deaths: number;
-  canFire: boolean;
-  isWalking: boolean;
-  playerSprite: Sprite | null = null;
-  playerSpriteLeft: Sprite | null = null;
-  interpolation: Vec2d | null = null;
-
-  constructor(
-    id: string,
-    x: number,
-    nickname: string,
-    santaColor: SantaColor,
-    y: number,
-    isLeft: boolean,
-    kills: number,
-    deaths: number,
-    canFire: boolean,
-    isWalking: boolean = false
-  ) {
-    this.id = id;
-    this.x = x;
-    this.nickname = nickname;
-    this.santaColor = santaColor;
-    this.y = y;
-    this.isLeft = isLeft;
-    this.kills = kills;
-    this.deaths = deaths;
-    this.canFire = canFire;
-    this.isWalking = isWalking;
-    this.loadSprites();
-  }
-
-  loadSprites() {
-    // TODO: Should be just one sprite with left and right frames on there,
-    // but I'm just going to do this for now
-    const image = new Image();
-    image.src = SANTA_SPRITE;
-    this.playerSprite = new Sprite(image, { x: 8, y: 6 }, { x: 32, y: 32 }, 0, {
-      x: this.x,
-      y: this.y,
-    });
-    this.playerSprite.buildFrameMap();
-    const imageLeft = new Image();
-    imageLeft.src = SANTA_SPRITE_LEFT;
-    this.playerSpriteLeft = new Sprite(
-      imageLeft,
-      { x: 8, y: 6 },
-      { x: 32, y: 32 },
-      0,
-      { x: this.x, y: this.y }
-    );
-    this.playerSpriteLeft.buildFrameMap();
-  }
-
-  updatePlayerInterpolation(interpolationFactor: number, maxIntDist: number) {
-    const startX = this.interpolation ? this.interpolation.x : this.x;
-    const startY = this.interpolation ? this.interpolation.y : this.y;
-
-    this.interpolation = {
-      x:
-        Math.abs(this.x - startX) > maxIntDist
-          ? this.x
-          : startX + interpolationFactor * (this.x - startX),
-      y:
-        Math.abs(this.y - startY) > maxIntDist
-          ? this.y
-          : startY + interpolationFactor * (this.y - startY),
-    };
-  }
-
-  draw(
-    ctx: CanvasRenderingContext2D,
-    cameraX: number,
-    cameraY: number,
-    isMe: boolean = false
-  ) {
-    // draw player
-    const { x, y } = this.interpolation ?? this;
-    const frameColorOffset =
-      (SantaColorToSpriteRow.get(this.santaColor) ?? 0) * 8;
-    // TODO: will be one sprite with left and right frames eventually
-    if (this.isLeft) {
-      this.playerSpriteLeft?.draw(
-        ctx,
-        x - cameraX,
-        y - cameraY,
-        frameColorOffset
-      );
-    } else {
-      this.playerSprite?.draw(ctx, x - cameraX, y - cameraY, frameColorOffset);
-    }
-    // draw name
-    let label = this.nickname;
-    ctx.fillStyle = "#00ff00";
-    if (isMe) {
-      label = "You";
-      ctx.fillStyle = "#ff0000";
-    }
-    ctx.font = "16px Arial";
-    // The "10"s are just offsets to make the label look better imo
-    ctx.fillText(
-      label,
-      x - cameraX - ctx.measureText(label).width / 2 + 10,
-      y - cameraY - 10
-    );
-  }
-}
+import Player from "./player/player";
 
 type Snowball = {
   id: number;
@@ -258,9 +68,6 @@ export async function start({
 
   const crosshairArmed = new Image();
   crosshairArmed.src = "/crosshair-armed.png";
-
-  const santaLeftImage = new Image();
-  const walkSnow = new Audio("walk-snow.mp3");
 
   const canvasEl = document.getElementById("canvas") as HTMLCanvasElement;
 
@@ -339,6 +146,7 @@ export async function start({
         players[idx].deaths = serverPlayer.deaths ?? players[idx].deaths;
         players[idx].canFire = serverPlayer.canFire ?? players[idx].canFire;
         players[idx].isLeft = serverPlayer.isLeft ?? players[idx].isLeft;
+        players[idx].isWalking = serverPlayer.isWalking ?? players[idx].isWalking;
       } else {
         let player = new Player(
           serverPlayer.id,
@@ -349,7 +157,8 @@ export async function start({
           serverPlayer.isLeft,
           serverPlayer.kills,
           serverPlayer.deaths,
-          serverPlayer.canFire
+          serverPlayer.canFire,
+          serverPlayer.isWalking
         );
         players.push(player);
       }
@@ -385,7 +194,6 @@ export async function start({
     const direction = keyDirectionMap.get(e.code);
     if (direction) {
       currentMoveDirection |= direction;
-      // walkSnow.play();
     }
     socket.emit("inputs", currentMoveDirection);
   });
@@ -399,8 +207,6 @@ export async function start({
     const direction = keyDirectionMap.get(e.code);
     if (direction) {
       currentMoveDirection &= ~direction;
-      walkSnow.pause();
-      walkSnow.currentTime = 0;
     }
     socket.emit("inputs", currentMoveDirection);
   });
@@ -497,7 +303,13 @@ export async function start({
     }
 
     for (const player of players) {
-      player.draw(canvas, cameraX, cameraY, getMyPlayer()?.id === player.id);
+      player.draw(
+        canvas,
+        cameraX,
+        cameraY,
+        delta,
+        getMyPlayer()?.id === player.id
+      );
     }
 
     for (const snowball of snowballs) {
